ch4/threads: added a new thead safety model: csync
CSync is meant for "combining synchronization". This is the
implementation of traditional software combining where a thread
announces a work request under contention and then waits for either the
lock to be release or the work has completed (executed by the combiner
thread on its behalf). The current prototype has been tested with
DSM-Sync, which is a software combining extension of the MCS lock.